requirements.txt
asyncpg==0.30.0
aiogram==3.19.0
APScheduler==3.11.0
python-decouple==3.8
python-dotenv==1.1.0
Flask==3.1.0
openpyxl==3.1.5
pandas==2.2.3
torch==2.7.0
sentence-transformers==4.1.0
hf-xet==1.1.2
numpy==2.0.2
pytest==8.4.0
pytest-asyncio==1.0.0